Добавил базу данных с использованием SQLite Api. Используется SQLite,Assets,NavComponent,SharedPreferences,Coroutine,Flows.

+++ app/src/main/java/com/bignerdranch/android/navigationcomponenttabs/screens/main/tabs/dashboard/BoxFragment.kt
@@ -43,7 +43,7 @@ class BoxFragment : Fragment(R.layout.fragment_box) {
         }
     }
 
-    private fun getBoxId(): Int = args.boxId
+    private fun getBoxId(): Long = args.boxId
 
     private fun getColorValue(): Int = args.colorValue
